Make header abstract
header is not abstract so we can also work with overviews which have a
subset of the header fields but can be much more quickly fetched.

Get rid of ThreadedHeader and just add children to Header. If we don't
want to persist is we'll just skip it. Simplifies the class hierarchy.