Allow spaces in quoted prefix
It looks like some postings end up with a space in the quoting string
so allow it and treat it as part of the prefix.

I'm sure there will be some degenerate cases where this isn't the right
thing to do but it seems to work a lot better for most of the cases that
I've seen.